chore: remove meteor.startup from jump-to-search-message
