+++ apps/meteor/client/startup/actionButtons/index.ts
@@ -1,6 +1,5 @@
 import './jumpToMessage';
 import './jumpToPinMessage';
-import './jumpToSearchMessage';
 import './jumpToStarMessage';
 import './permalinkPinned';
 import './permalinkStar';
