--- apps/meteor/client/startup/actionButtons/jumpToSearchMessage.ts
@@ -1,18 +1,0 @@
-import { Meteor } from 'meteor/meteor';
-
-import { MessageAction } from '../../../app/ui-utils/client';
-import { setMessageJumpQueryStringParameter } from '../../lib/utils/setMessageJumpQueryStringParameter';
-
-Meteor.startup(() => {
-	MessageAction.addButton({
-		id: 'jump-to-search-message',
-		icon: 'jump',
-		label: 'Jump_to_message',
-		context: ['search'],
-		async action(_, { message }) {
-			setMessageJumpQueryStringParameter(message._id);
-		},
-		order: 100,
-		group: 'message',
-	});
-});
